backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
scikit-learn>=1.5.0
PyPDF2>=3.0.0
python-docx>=1.1.0
